Four is not prime - refactor
I make explicit that the new condition that appeared in the code
is actually our check for primality. When we'll jump into the real
'criba' algorithm this condition will be checking for marked
numbers.

Since we know we want to implement the 'criba' algorithm, I'll
start refactoring now to adapt to this algorithm.