Eight is not prime - refactor
Finally we have a 'criba', an array of marked numbers, which we
called number_marks. It's just a cache of the former isPrime
method.

Of course, it's an annoying structure since adding new test cases
forces us to change the implementation still adding cases to the
checkIfNumberIsPrime method, but now that we already have two
loops we can easily use the 'criba' method to replace that
hard-coded primality function.